small parser mods + move dict_merge into utils
